amd64: Make the port compile and boot again.
This also fixes a lot of serious bugs, like picking a random end address. It was not actually used as we always mapped a fixed amount of pages, but this commit fixes that as well.

+++ include/sys/amd64/_types.h
@@ -11,8 +11,8 @@ typedef short		int16_t;
 typedef unsigned short	uint16_t;
 typedef int		int32_t;
 typedef unsigned int	uint32_t;
-typedef long long	int64_t;
-typedef unsigned long long uint64_t;
+typedef long		int64_t;
+typedef unsigned long	uint64_t;
 
 /* 7.18.1.2: Minimum-width integer types */
 typedef uint8_t		uint_least8_t;
